delete attack interfaces
