+++ src/battle/attack/Attack.java
@@ -3,7 +3,6 @@ package battle.attack;
 import battle.Battle;
 import battle.effect.PassableEffect;
 import battle.effect.SapHealthEffect;
-import battle.effect.attack.AttackInterfaces.RecoilPercentageMove;
 import battle.effect.attack.ChangeAbilityMove;
 import battle.effect.attack.ChangeTypeSource;
 import battle.effect.attack.MultiStrikeMove;
@@ -30,6 +29,7 @@ import battle.effect.generic.EffectInterfaces.OpponentStatSwitchingEffect;
 import battle.effect.generic.EffectInterfaces.PhysicalContactEffect;
 import battle.effect.generic.EffectInterfaces.RapidSpinRelease;
 import battle.effect.generic.EffectInterfaces.RecoilMove;
+import battle.effect.generic.EffectInterfaces.RecoilPercentageMove;
 import battle.effect.generic.EffectInterfaces.SelfAttackBlocker;
 import battle.effect.generic.EffectInterfaces.SleepyFightsterEffect;
 import battle.effect.generic.EffectInterfaces.SwapOpponentEffect;
