+++ src/battle/effect/generic/EffectInterfaces.java
@@ -8,9 +8,12 @@ import battle.effect.status.StatusCondition;
 import item.Item;
 import main.Global;
 import map.TerrainType;
+import message.MessageUpdate;
+import message.Messages;
 import pokemon.ActivePokemon;
 import pokemon.Stat;
 import pokemon.ability.Ability;
+import pokemon.ability.AbilityNamesies;
 import trainer.Trainer;
 import type.Type;
 
@@ -105,6 +108,19 @@ public final class EffectInterfaces {
 					effect.applyRecoil(b, user, damage);
 				}
 			}
+		}
+	}
+
+	public interface RecoilPercentageMove extends RecoilMove {
+		int getDamagePercentageDenominator();
+
+		default void applyRecoil(Battle b, ActivePokemon user, int damage) {
+			if (user.hasAbility(AbilityNamesies.ROCK_HEAD) || user.hasAbility(AbilityNamesies.MAGIC_GUARD)) {
+				return;
+			}
+			
+			Messages.add(new MessageUpdate(user.getName() + " was hurt by recoil!"));
+			user.reduceHealth(b, (int)Math.ceil((double)damage/getDamagePercentageDenominator()), false);
 		}
 	}
